refactor(contentful-image): rename component
