+++ atomic-design/atoms/contentful-image/__tests__/contentful-image.tsx
@@ -1,39 +1,39 @@
 import { createTree, Wrapper } from "@evernest/dev-helpers";
 import { mount } from "enzyme";
 import React from "react";
-import { ResponsiveImage } from "../src";
+import { ContentfulImage } from "../src";
 
-const image = {
-	url:
+const imageArgs = {
+	src:
 		"//images.ctfassets.net/vij0xfkj6nfm/65B2DBSyhcTksD4H9nuGds/a3be81862610240ee35085e63835f531/007-web-1000px.jpg",
 	width: 1000,
 	height: 667,
-	title: "placeholder image",
+	alt: "placeholder image",
 };
 
-test("ResponsiveImage renders", () => {
+test("ContentfulImage renders", () => {
 	const tree = createTree(
 		<Wrapper>
-			<ResponsiveImage image={image} />
+			<ContentfulImage {...imageArgs} />
 		</Wrapper>
 	);
 	expect(tree).toMatchSnapshot();
 });
 
-test("ResponsiveImage mounts", () => {
+test("ContentfulImage mounts", () => {
 	const wrapper = mount(
 		<Wrapper>
-			<ResponsiveImage image={image} />
+			<ContentfulImage {...imageArgs} />
 		</Wrapper>
 	);
-	expect(wrapper.find(ResponsiveImage));
+	expect(wrapper.find(ContentfulImage));
 });
 
-test("ResponsiveImage calls onLoad", () => {
+test("ContentfulImage calls onLoad", () => {
 	const callback = jest.fn(() => {});
 	const wrapper = mount(
 		<Wrapper>
-			<ResponsiveImage image={image} onLoad={callback} />
+			<ContentfulImage {...imageArgs} onLoad={callback} />
 		</Wrapper>
 	);
 	wrapper.find("img").at(1).simulate("load");
