+++ atomic-design/atoms/contentful-image/index.js
