+++ atomic-design/atoms/contentful-image/jest.config.js
