+++ atomic-design/atoms/contentful-image/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@evernest/responsive-image",
+	"name": "@evernest/contentful-image",
 	"version": "0.0.0",
 	"keywords": [
 		"a11y",
