+++ atomic-design/atoms/contentful-image/rollup.config.js
