+++ atomic-design/atoms/contentful-image/src/index.ts
@@ -1,3 +1,3 @@
-export * from "./responsive-image";
+export * from "./contentful-image";
 export * from "./utils";
 export * from "./types";
