+++ atomic-design/atoms/contentful-image/src/utils.ts
