+++ atomic-design/atoms/contentful-image/tsconfig.json
